fix: wire agent registry through AgentLifecyclePort (closes #94)
- Change CliState.agent_registry to Arc<dyn AgentLifecyclePort>
- Create src/domain/agent.rs for AgentEntry/AgentMetadata types
- Update agent handlers to use port trait
