+++ src/domain/mod.rs
@@ -1,3 +1,4 @@
+pub mod agent;
 pub mod belief;
 pub mod memory;
 pub mod pattern;
